pyproject.toml
[project]
name = "parking-spot-monitor"
version = "0.1.0"
description = "Local parking spot monitor for UniFi Protect RTSP streams."
readme = "parking-spot-monitor-spec.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "Pillow>=10",
    "pydantic>=2",
    "PyYAML>=6",
    "ultralytics>=8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
